Promote cell count to long
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@58118 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ raster/r.random/count.c
@@ -55,7 +55,7 @@ void get_stats(struct rr_state *theState)
     nrows = Rast_window_rows();
     ncols = Rast_window_cols();
 
-    theState->nCells = nrows * ncols;
+    theState->nCells = (long) nrows * ncols;
     theState->nNulls = 0;
     set_min(NULL, 0, &theState->min);
     set_max(NULL, 0, &theState->max);
